Solve assignment in Zote and Python
This was a great assignment about introductory indexing (for databases),
given at the fourth summer school in data science given by ACM Europe.
It asked you to do a range and a k-NN query for some 10D data. First
through the naive way, and then using a pivot point idea utilizing the
triangle inequality to avoid doing as many real distance computations.
The assumpion was that the distance would be expensive, when compared
to the rest.

I solved it during the school in Python, but then went back and did it
in my own language Zote as well. In my opinion the Zote code is really
clean, but it also takes around 3-10 times longer than the Python code
to execute (which is reasonable as its implementation is just an
interpreter over the AST, and not optimized at all yet).